refactor: clarify channel input boundary
