+++ src/runtime/channel_input.rs
@@ -1,4 +1,4 @@
-//! Runtime ingress helpers for channel-originated turns.
+//! Runtime boundary for channel-originated input.
 //!
 //! Channels normalize platform events first; this module translates those
 //! normalized inputs into EgoPulse runtime work: `SurfaceContext`, Channel Log
@@ -119,5 +119,74 @@ pub(super) fn submit_scheduled_turn(state: &AppState, scheduled: ScheduledTurn) 
         tokio::spawn(async move {
             crate::runtime::execute_scheduled_turn(&state, turn).await;
         });
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use crate::runtime::build_sleep_app_state_with_path;
+
+    fn build_test_state(dir: &tempfile::TempDir) -> AppState {
+        let config = crate::test_util::test_config(dir.path().to_str().expect("utf8"));
+        build_sleep_app_state_with_path(config, Some(dir.path().join("egopulse.config.yaml")))
+            .expect("build sleep state")
+    }
+
+    #[test]
+    fn channel_scope_from_secret_maps_channel_flag() {
+        assert_eq!(channel_scope_from_secret(true), ConversationScope::Secret);
+        assert_eq!(channel_scope_from_secret(false), ConversationScope::Normal);
+    }
+
+    #[test]
+    fn build_channel_context_applies_scope_and_session_identity() {
+        let context = build_channel_context(
+            "discord",
+            "alice",
+            "123",
+            "discord",
+            "reviewer",
+            ConversationScope::Secret,
+        );
+
+        assert_eq!(context.channel, "discord");
+        assert_eq!(context.surface_user, "alice");
+        assert_eq!(context.surface_thread, "123");
+        assert_eq!(context.chat_type, "discord");
+        assert_eq!(context.agent_id, "reviewer");
+        assert_eq!(context.scope, ConversationScope::Secret);
+        assert_eq!(context.session_key(), "discord:123:agent:reviewer");
+    }
+
+    #[tokio::test]
+    async fn store_human_channel_log_message_persists_discord_message() {
+        let dir = tempfile::tempdir().expect("tempdir");
+        let state = build_test_state(&dir);
+
+        let chat_id = store_human_channel_log_message(
+            &state,
+            HumanChannelLogMessage {
+                key: ChannelLogKey::Discord(42),
+                scope: ConversationScope::Normal,
+                id: "cl-42".to_string(),
+                sender_id: "user:discord:7".to_string(),
+                content: "hello".to_string(),
+                timestamp: "2026-06-25T00:00:00Z".to_string(),
+            },
+        )
+        .await
+        .expect("channel log chat id");
+
+        let messages = state
+            .db_for(ConversationScope::Normal)
+            .get_channel_log_messages(chat_id, 10)
+            .expect("messages");
+        assert_eq!(messages.len(), 1);
+        assert_eq!(messages[0].id, "cl-42");
+        assert_eq!(messages[0].sender_id, "user:discord:7");
+        assert_eq!(messages[0].sender_kind, SenderKind::User);
+        assert_eq!(messages[0].content, "hello");
     }
 }
